refactor order page
