--- src/components/pages/index.ts
@@ -1,2 +1,0 @@
-export * from './order/table'
-export * from './order/calculate-total'
